Moved settings in configuration
